30 de octubre de 2020

+++ src/main/java/org/uma/mbd/mdBancoV1L/banco/Cuenta.java
@@ -31,7 +31,7 @@ public class Cuenta {
         saldo += cant;
     }
 
-    public void debito(int cant) {
+    public void debito(double cant) {
         saldo -= cant;
     }
